backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.40
asyncpg==0.30.0
alembic==1.15.2
pydantic[email]==2.11.3
pydantic-settings==2.9.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
python-dotenv==1.1.0
httpx==0.28.1
# S3を使う場合のみ必要
boto3==1.38.0
